dsp_uas_model/requirements.txt
flask==3.1.2
joblib==1.3.2
mlflow==2.12.1
python-dotenv==1.0.1
requests==2.31.0
blinker==1.9.0
click==8.1.8
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
setuptools==68.0.0
gunicorn==20.1.0
dagshub